external_interrupt works nrf_dfu_extFlash_rpc and nrf_dfu_intFlash_rpc work in ncs v2.0.0

+++ ble_rpc/nrf_dfu_extFlash_rpc/src/rpc_app_api.h
@@ -22,7 +22,7 @@ extern "C" {
 #endif
 
 int app2net_send_nus(uint8_t *data, uint16_t len);
-int rpc_app2net_send(int type, uint8_t *buffer, size_t length);
+int rpc_app2net_send(int32_t type, uint8_t *buffer, int32_t length);
 int net2app_receive(int type, uint8_t *data, size_t len);
 bool is_ble_connected(void);
